# New Performance/DoubleStartEndWith cop