This is a continuation of https://github.com/bbatsov/rubocop/pull/2535 .

This cop checks for double `#start_with?` or `#end_with?` calls
separated by `||`. In some cases such calls can be replaced
with an single `#start_with?`/`#end_with?` call.

```ruby
  # bad
  str.start_with?("a") || str.start_with?(Some::CONST)
  str.start_with?("a", "b") || str.start_with?("c")
  var1 = ...
  var2 = ...
  str.end_with?(var1) || str.end_with?(var2)

  # good
  str.start_with?("a", Some::CONST)
  str.start_with?("a", "b", "c")
  var1 = ...
  var2 = ...
  str.end_with?(var1, var2)
```

The corner cases are when arguments to start_with? have side effects,
or when the receiver itself has side effects, or when the receiver
is not a string at all, or when someone monkey-patched `String#start_with?`

Benchmark results are not too convincing because of a big variance, but they show that the single-call approach is at least not slower that the two-call approach.

```ruby
prefix1 = "lib/"
prefix2 = "test/"
{
  'First prefix matches' => 'lib/file.rb',
  'Second prefix matches' => 'test/file.rb',
  'None matches' => 'app/models/file.rb'
}.each do |description, str|
  Benchmark.ips do |x|
    puts "==== #{description} ===="
    x.report('two calls') { str.start_with?(prefix1) || str.start_with?(prefix2) }
    x.report('one call')  { str.start_with?(prefix1, prefix2) }
    x.compare!
  end
end
```

```
==== First prefix matches ====
Calculating -------------------------------------
           two calls    31.010k i/100ms
            one call    35.180k i/100ms
-------------------------------------------------
           two calls      2.035M (±14.6%) i/s -      9.644M
            one call      2.039M (±15.9%) i/s -      9.569M

Comparison:
            one call:  2038728.5 i/s
           two calls:  2034701.2 i/s - 1.00x slower

==== Second prefix matches ====
Calculating -------------------------------------
           two calls    31.139k i/100ms
            one call    34.518k i/100ms
-------------------------------------------------
           two calls      1.515M (±20.7%) i/s -      6.944M
            one call      1.709M (±19.5%) i/s -      7.836M

Comparison:
            one call:  1708570.7 i/s
           two calls:  1514975.3 i/s - 1.13x slower

==== None matches ====
Calculating -------------------------------------
           two calls    33.038k i/100ms
            one call    36.789k i/100ms
-------------------------------------------------
           two calls      1.471M (±18.9%) i/s -      6.773M
            one call      1.778M (±19.9%) i/s -      8.130M

Comparison:
            one call:  1778026.2 i/s
           two calls:  1471409.3 i/s - 1.21x slower
```

Wild life occurrences: https://github.com/ManageIQ/manageiq/pull/5558, https://github.com/rails/rails/pull/22374, the aforementioned https://github.com/bbatsov/rubocop/pull/2535.