@alexdowad 2016-01-06
Nice!! I was thinking of doing this myself, good to see that someone else did it first!