@alexdowad 2016-01-06
One other way this could be improved -- think of things like this:

```ruby
# this will parse as (or (or something str.start_with?('a')) str.start_with?('b'))...
# in other words the 2 nodes you are interested in will not be nested under the same 'or' node
something || str.start_with?('a') || str.start_with?('b')

# another example:
str.start_with?('a') || foo || bar || str.start_with?('b')
```

Evaluation is from left to right, so you have to make sure that there is no node with unknown side effects between the 2 matching calls.