@DNNX 2016-01-06
`something || str.start_with?('a') || str.start_with?('b')` - yea, I was thinking about this, but found it kind of complex to implement so I decided to go with more simple implementation which covers most real-life cases (in my experience). I also saw `!str.start_with?(a) && !str.start_with?(b)` in one open-source project. I can also be simplified, but again, I decided not to go this far. Do you think it should be done in this PR?

Another thing which can potentially be done is we can loosen the requirement of the first argument to the first `start_with?` call to be pure. As long as all other arguments are pure, the first one can have side effects  - the semantics will stay the same.