@alexdowad 2016-01-06
> Do you think it should be done in this PR?

It's up to you. I am just suggesting ways to make the PR better. If you don't want to go that far in this PR, and the maintainers want to accept it as is, that is fine. It is already an improvement over what we have.