@DNNX 2016-01-06
Here is a little summary:

- [x] Implemented more intelligent matcher as per @alexdowad's suggestion
- [x] Improved offense message format. No more abstract `x, ... y, ...`; the message now includes the actual code
- [x] Dropped the requirement for the arguments of the first `start_with?` call to be pure
- [ ] `something || str.start_with?('a') || str.start_with?('b')` - not going to do this in this PR if you guys don't mind
- [ ] `str.start_with?('a') || foo || bar || str.start_with?('b')` - will leave it as well
- [ ] `!str.start_with?(a) && !str.start_with?(b)` - not going to do this either (in this PR at least)

If you guys are ok with this, please let me know. I will squash the commits and add an entry to the change log then.