@DNNX 2016-01-07
UPD: This PR is ready to merge now, if everyone is ok with the changes.