@bbatsov 2016-01-07
Rebase this and I'll have it merged.