@DNNX 2016-01-07
Rebased it. CI green, no conflicts.